+++ cocos2d/cocos/2d/CCLabelBMFont.h
@@ -125,7 +125,7 @@ public:
     virtual const BlendFunc &getBlendFunc() const override;
 
     virtual Sprite * getLetter(int ID);
-    virtual Node * getChildByTag(int tag) const override;
+    virtual Node * getChildByTag(int64_t tag) const override;
 
     virtual void setColor(const Color3B& color) override;
