+++ cocos2d/cocos/2d/CCMenu.h
@@ -167,7 +167,7 @@ public:
     
     virtual void addChild(Node * child) override;
     virtual void addChild(Node * child, int zOrder) override;
-    virtual void addChild(Node * child, int zOrder, int tag) override;
+    virtual void addChild(Node * child, int zOrder, int64_t tag) override;
     virtual void addChild(Node * child, int zOrder, const std::string &name) override;
     
     virtual void onEnter() override;
