+++ cocos2d/cocos/2d/CCParallaxNode.h
@@ -82,7 +82,7 @@ public:
     //
     // Overrides
     //
-    virtual void addChild(Node * child, int zOrder, int tag) override;
+    virtual void addChild(Node * child, int zOrder, int64_t tag) override;
     virtual void addChild(Node * child, int zOrder, const std::string &name) override;
     virtual void removeChild(Node* child, bool cleanup) override;
     virtual void removeAllChildrenWithCleanup(bool cleanup) override;
