+++ cocos2d/cocos/ui/UIScrollView.h
@@ -376,7 +376,7 @@ public:
     //override functions
     virtual void addChild(Node* child)override;
     virtual void addChild(Node * child, int localZOrder)override;
-    virtual void addChild(Node* child, int localZOrder, int tag) override;
+    virtual void addChild(Node* child, int localZOrder, int64_t tag) override;
     virtual void addChild(Node* child, int localZOrder, const std::string &name) override;
     virtual void removeAllChildren() override;
     virtual void removeAllChildrenWithCleanup(bool cleanup) override;
@@ -384,7 +384,7 @@ public:
     virtual Vector<Node*>& getChildren() override;
     virtual const Vector<Node*>& getChildren() const override;
     virtual ssize_t getChildrenCount() const override;
-    virtual Node * getChildByTag(int tag) const override;
+    virtual Node * getChildByTag(int64_t tag) const override;
     virtual Node* getChildByName(const std::string& name)const override;
     //touch event callback
     virtual bool onTouchBegan(Touch *touch, Event *unusedEvent) override;
